refactor(notifications): Step D — extract @sip-and-speak/notifications package
Lifts the notification dispatcher out of apps/server into a new workspace
package. apps/server now imports a single function — the dispatcher's
internals are no longer reachable from outside the package.

Moves:
- apps/server/src/notification-delivery.ts → packages/notifications/src/delivery.ts
- apps/server/src/notification-builders.ts → packages/notifications/src/builders.ts
- apps/server/src/notification-recipe.ts → packages/notifications/src/recipe.ts
- apps/server/src/notifications.ts → packages/notifications/src/dispatcher.ts
- apps/server/src/__tests__/notifications-*.test.ts → packages/notifications/src/__tests__/

Adds:
- packages/notifications/package.json + tsconfig + index.ts (public surface)
- packages/notifications/bunfig.toml + test-preload.ts to seed env vars in tests
  (apps/server's .env was supplying them implicitly pre-move)

Kills the cross-package leak: notification body composition
(buildMatchRequestNotificationBody) is now a private helper in builders.ts
instead of an export from packages/api/src/routers/matching-utils.ts. The
4 body-composition tests that used to live in packages/api/src/__tests__/
matching.test.ts move to packages/notifications/src/__tests__/builders.test.ts
(expanded — pure tests, no mocks).

apps/server/src/index.ts: imports registerNotificationHandlers from
@sip-and-speak/notifications. apps/server/package.json adds the workspace
dependency.

Test totals across the workspace remain healthy (existing fragility
unchanged; 4 new pure body tests added in the right home). Type errors
unchanged (10 total split as 4 in notifications + 6 in api/chat.ts).
