+++ packages/notifications/src/__tests__/notifications-meetup-cancelled.test.ts
@@ -25,7 +25,7 @@ mock.module("@sip-and-speak/db", () => ({
   },
 }));
 
-import { registerNotificationHandlers } from "../notifications";
+import { registerNotificationHandlers } from "../dispatcher";
 import { domainEvents } from "@sip-and-speak/api/domain-events";
 
 describe("handleMeetupCancelled", () => {
