+++ packages/notifications/src/__tests__/notifications-meetup-confirmed.test.ts
@@ -1,8 +1,8 @@
 /**
- * Tests for task #77 — Push notification on MeetupDeclined
+ * Tests for task #75 — Push notification on MeetupConfirmed
  *
  * Covers:
- *   - Both Students notified when a proposal is declined
+ *   - Both Students notified when a meetup is confirmed
  *   - No notification sent when neither Student has a device token
  */
 import { describe, it, expect, mock, beforeEach } from "bun:test";
@@ -28,8 +28,10 @@ const fetchCalls: CapturedFetchCall[] = [];
 
 // ── Schema mocks ──────────────────────────────────────────────────────────────
 
+const DEVICE_TOKEN_TABLE = "userDeviceToken";
+
 mock.module("@sip-and-speak/db/schema/sip-and-speak", () => ({
-  userDeviceToken: "userDeviceToken",
+  userDeviceToken: DEVICE_TOKEN_TABLE,
   userLanguage: "userLanguage",
 }));
 
@@ -45,16 +47,19 @@ mock.module("drizzle-orm", () => ({
 
 // ── DB mock ───────────────────────────────────────────────────────────────────
 
+// Track which userId is queried to return the right tokens
 let proposerTokens: Array<{ id: string; token: string }> = [];
 let receiverTokens: Array<{ id: string; token: string }> = [];
-const PROPOSER_ID = "proposer-abc";
-const RECEIVER_ID = "receiver-xyz";
+let lastQueriedUserId = "";
+const PROPOSER_ID = "proposer-123";
+const RECEIVER_ID = "receiver-456";
 
 mock.module("@sip-and-speak/db", () => ({
   db: {
     select: () => ({
       from: () => ({
         where: (clause: { _val: string }) => {
+          lastQueriedUserId = clause._val;
           const tokens = clause._val === PROPOSER_ID ? proposerTokens : receiverTokens;
           return Promise.resolve(tokens);
         },
@@ -65,10 +70,10 @@ mock.module("@sip-and-speak/db", () => ({
 
 // ── Subject under test ────────────────────────────────────────────────────────
 
-import { registerNotificationHandlers } from "../notifications";
+import { registerNotificationHandlers } from "../dispatcher";
 import { domainEvents } from "@sip-and-speak/api/domain-events";
 
-describe("handleMeetupDeclined", () => {
+describe("handleMeetupConfirmed", () => {
   beforeEach(() => {
     fetchCalls.length = 0;
     proposerTokens = [];
@@ -76,36 +81,44 @@ describe("handleMeetupDeclined", () => {
     registerNotificationHandlers();
   });
 
-  it("notifies both Students when a proposal is declined", async () => {
+  it("notifies both Students when a meetup is confirmed", async () => {
     proposerTokens = [{ id: "pt-1", token: "ExponentPushToken[proposer-token]" }];
     receiverTokens = [{ id: "rt-1", token: "ExponentPushToken[receiver-token]" }];
 
-    domainEvents.emit("MeetupDeclined", {
+    domainEvents.emit("MeetupConfirmed", {
       meetupId: "meetup-1",
       proposerId: PROPOSER_ID,
       receiverId: RECEIVER_ID,
-      declinedAt: new Date(),
+      venueName: "Atlas Building",
+      date: "2026-05-10",
+      time: "14:00",
+      confirmedAt: new Date(),
     });
 
+    // Allow async handler to flush
     await new Promise((r) => setTimeout(r, 20));
 
     expect(fetchCalls.length).toBe(1);
-    const msgs = fetchCalls[0]!.messages;
-    expect(msgs.length).toBe(2);
-    expect(msgs.map((m) => m.to)).toContain("ExponentPushToken[proposer-token]");
-    expect(msgs.map((m) => m.to)).toContain("ExponentPushToken[receiver-token]");
-    expect(msgs[0]!.title).toBe("Meetup proposal declined");
+    const sentMessages = fetchCalls[0]!.messages;
+    expect(sentMessages.length).toBe(2);
+    expect(sentMessages.map((m) => m.to)).toContain("ExponentPushToken[proposer-token]");
+    expect(sentMessages.map((m) => m.to)).toContain("ExponentPushToken[receiver-token]");
+    expect(sentMessages[0]!.title).toBe("Meetup confirmed! 🎉");
+    expect(sentMessages[0]!.body).toContain("Atlas Building");
   });
 
   it("sends no notification when neither Student has a device token", async () => {
     proposerTokens = [];
     receiverTokens = [];
 
-    domainEvents.emit("MeetupDeclined", {
+    domainEvents.emit("MeetupConfirmed", {
       meetupId: "meetup-2",
       proposerId: PROPOSER_ID,
       receiverId: RECEIVER_ID,
-      declinedAt: new Date(),
+      venueName: "Flux",
+      date: "2026-05-11",
+      time: "10:00",
+      confirmedAt: new Date(),
     });
 
     await new Promise((r) => setTimeout(r, 20));
