+++ packages/notifications/src/__tests__/notifications-meetup-declined.test.ts
@@ -1,9 +1,9 @@
 /**
- * Tests for task #76 — Push notification on MeetupCounterProposed
+ * Tests for task #77 — Push notification on MeetupDeclined
  *
  * Covers:
- *   - Original proposer (now the new receiver) notified when a counter-proposal arrives
- *   - No notification sent when the new receiver has no device token
+ *   - Both Students notified when a proposal is declined
+ *   - No notification sent when neither Student has a device token
  */
 import { describe, it, expect, mock, beforeEach } from "bun:test";
 
@@ -45,14 +45,19 @@ mock.module("drizzle-orm", () => ({
 
 // ── DB mock ───────────────────────────────────────────────────────────────────
 
-let mockTokenRows: Array<{ id: string; token: string }> = [];
-const ORIGINAL_PROPOSER_ID = "orig-proposer-123"; // becomes newReceiverId after counter
+let proposerTokens: Array<{ id: string; token: string }> = [];
+let receiverTokens: Array<{ id: string; token: string }> = [];
+const PROPOSER_ID = "proposer-abc";
+const RECEIVER_ID = "receiver-xyz";
 
 mock.module("@sip-and-speak/db", () => ({
   db: {
     select: () => ({
       from: () => ({
-        where: () => Promise.resolve(mockTokenRows),
+        where: (clause: { _val: string }) => {
+          const tokens = clause._val === PROPOSER_ID ? proposerTokens : receiverTokens;
+          return Promise.resolve(tokens);
+        },
       }),
     }),
   },
@@ -60,51 +65,47 @@ mock.module("@sip-and-speak/db", () => ({
 
 // ── Subject under test ────────────────────────────────────────────────────────
 
-import { registerNotificationHandlers } from "../notifications";
+import { registerNotificationHandlers } from "../dispatcher";
 import { domainEvents } from "@sip-and-speak/api/domain-events";
 
-describe("handleMeetupCounterProposed", () => {
+describe("handleMeetupDeclined", () => {
   beforeEach(() => {
     fetchCalls.length = 0;
-    mockTokenRows = [];
+    proposerTokens = [];
+    receiverTokens = [];
     registerNotificationHandlers();
   });
 
-  it("notifies the original proposer (new receiver) of the counter-proposal", async () => {
-    mockTokenRows = [{ id: "token-1", token: "ExponentPushToken[orig-proposer]" }];
+  it("notifies both Students when a proposal is declined", async () => {
+    proposerTokens = [{ id: "pt-1", token: "ExponentPushToken[proposer-token]" }];
+    receiverTokens = [{ id: "rt-1", token: "ExponentPushToken[receiver-token]" }];
 
-    domainEvents.emit("MeetupCounterProposed", {
+    domainEvents.emit("MeetupDeclined", {
       meetupId: "meetup-1",
-      newProposerId: "counter-proposer-456",
-      newReceiverId: ORIGINAL_PROPOSER_ID,
-      venueName: "Vertigo",
-      date: "2026-05-15",
-      time: "10:30",
-      round: 2,
-      counterProposedAt: new Date(),
+      proposerId: PROPOSER_ID,
+      receiverId: RECEIVER_ID,
+      declinedAt: new Date(),
     });
 
     await new Promise((r) => setTimeout(r, 20));
 
     expect(fetchCalls.length).toBe(1);
-    const msg = fetchCalls[0]!.messages[0]!;
-    expect(msg.to).toBe("ExponentPushToken[orig-proposer]");
-    expect(msg.title).toContain("round 2");
-    expect(msg.body).toContain("Vertigo");
+    const msgs = fetchCalls[0]!.messages;
+    expect(msgs.length).toBe(2);
+    expect(msgs.map((m) => m.to)).toContain("ExponentPushToken[proposer-token]");
+    expect(msgs.map((m) => m.to)).toContain("ExponentPushToken[receiver-token]");
+    expect(msgs[0]!.title).toBe("Meetup proposal declined");
   });
 
-  it("sends no notification when new receiver has no device token", async () => {
-    mockTokenRows = [];
+  it("sends no notification when neither Student has a device token", async () => {
+    proposerTokens = [];
+    receiverTokens = [];
 
-    domainEvents.emit("MeetupCounterProposed", {
+    domainEvents.emit("MeetupDeclined", {
       meetupId: "meetup-2",
-      newProposerId: "counter-proposer-456",
-      newReceiverId: ORIGINAL_PROPOSER_ID,
-      venueName: "Atlas",
-      date: "2026-05-16",
-      time: "11:00",
-      round: 2,
-      counterProposedAt: new Date(),
+      proposerId: PROPOSER_ID,
+      receiverId: RECEIVER_ID,
+      declinedAt: new Date(),
     });
 
     await new Promise((r) => setTimeout(r, 20));
