+++ packages/notifications/src/__tests__/notifications-removal-conversations.test.ts
@@ -50,7 +50,7 @@ mock.module("@sip-and-speak/db", () => ({
   },
 }));
 
-import { registerNotificationHandlers } from "../notifications";
+import { registerNotificationHandlers } from "../dispatcher";
 import { domainEvents } from "@sip-and-speak/api/domain-events";
 
 describe("handleStudentRemovedCloseConversations (#111)", () => {
