+++ packages/notifications/src/__tests__/notifications-removal.test.ts
@@ -59,7 +59,7 @@ mock.module("@sip-and-speak/db", () => ({
   },
 }));
 
-import { registerNotificationHandlers } from "../notifications";
+import { registerNotificationHandlers } from "../dispatcher";
 import { domainEvents } from "@sip-and-speak/api/domain-events";
 
 describe("handleStudentRemovedNotify (#112)", () => {
