+++ packages/notifications/src/__tests__/notifications-suspension.test.ts
@@ -70,7 +70,7 @@ mock.module("@sip-and-speak/db", () => ({
   },
 }));
 
-import { registerNotificationHandlers } from "../notifications";
+import { registerNotificationHandlers } from "../dispatcher";
 import { domainEvents } from "@sip-and-speak/api/domain-events";
 
 describe("handleStudentSuspended — proposal cancellation (#102)", () => {
